Rename: PostDetail관련 스타일 파일 합치기
PostHeader, PostBody, PostFooter .style -> PostDetail.style
