+++ src/components/PostDetail/PostBody/index.tsx
@@ -1,4 +1,4 @@
-import * as S from './PostBody.style'
+import * as S from '../PostDetail.style'
 import PostContent from './PostContent'
 import TableOfContent from './TableOfContent'
 
@@ -10,10 +10,10 @@ interface PostBodyProps {
 // 문자열 형태의 HTML 코드를 출력하는 것도 dangerouslySetInnerHTML 속성을 통해 간단하게 구현이 가능
 const PostBody = ({ tableOfContents, html }: PostBodyProps) => {
   return (
-    <S.Container>
+    <S.Body>
       <TableOfContent tableOfContents={tableOfContents} />
       <PostContent html={html} />
-    </S.Container>
+    </S.Body>
   )
 }
