+++ src/components/PostDetail/PostDetail.style.tsx
@@ -3,19 +3,20 @@ import { GatsbyImage } from 'gatsby-plugin-image'
 
 import { GatsbyImageDataType } from '@/types/gatsby.type'
 
+// Header
 interface GatsbyImgProps {
   image: GatsbyImageDataType
   alt: string
   className?: string
 }
 
-export const PostHeadWrapper = styled.div`
+export const Header = styled.div`
   position: relative;
   width: 100%;
   height: 400px;
   margin-bottom: 64px;
 
-  @media (max-width: 768px) {
+  @media ${({ theme }) => theme.media.medium} {
     height: 300px;
   }
 `
@@ -23,19 +24,14 @@ export const PostHeadWrapper = styled.div`
 export const PostHeadInfoWrapper = styled.div`
   display: flex;
   flex-direction: column;
-  width: 768px;
+  width: var(--main-content-width);
   height: 100%;
   margin: 0 auto;
-  padding: 60px 0;
+  padding: var(--padding-l) var(--space-m);
   color: var(--color-white);
   background-color: transparent;
   position: relative;
-  z-index: 100;
-
-  @media (max-width: 768px) {
-    width: 100%;
-    padding: 40px 20px;
-  }
+  z-index: var(--z-index-second);
 `
 
 export const Title = styled.div`
@@ -47,10 +43,10 @@ export const Title = styled.div`
   white-space: normal;
   -webkit-line-clamp: 2;
   -webkit-box-orient: vertical;
-  ${({ theme: { typography } }) => typography.displayBoldLarge}
+  ${({ theme: { typography } }) => typography.displayBoldLarge};
 
-  @media (max-width: 768px) {
-    font-size: 30px;
+  @media ${({ theme }) => theme.media.medium} {
+    ${({ theme: { typography } }) => typography.displayBold};
   }
 `
 
@@ -62,11 +58,14 @@ export const PostData = styled.div`
 
   ${({ theme: { typography } }) => typography.linkMedium}
 
-  @media (max-width: 768px) {
+  @media ${({ theme: { media } }) => media.medium} {
     flex-direction: column;
     align-items: flex-start;
-    font-size: 15px;
-    font-weight: 400;
+    ${({ theme: { typography } }) => typography.linkBase}
+  }
+
+  @media ${({ theme: { media } }) => media.small} {
+    ${({ theme: { typography } }) => typography.linkSmall}
   }
 `
 
@@ -75,23 +74,72 @@ export const PostInfo = styled.div`
   gap: 8px;
 `
 
-export const PrevPageIcon = styled.div`
-  display: grid;
-  place-items: center;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  background-color: var(--color-white);
-  color: var(--color-black);
-  font-size: 22px;
-  cursor: pointer;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+// Body
+export const Body = styled.div`
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  width: var(--main-content-width);
+  margin: 0 auto;
+`
 
-  @media (max-width: 768px) {
-    width: 30px;
-    height: 30px;
-    font-size: 18px;
+export const MarkdownRenderer = styled.article`
+  ${({ theme: { markdownStyle } }) => markdownStyle}
+`
+
+export const TableOfContent = styled.aside`
+  position: absolute;
+  height: 100%;
+  right: -45%;
+  font-size: 0.9rem;
+
+  > ul {
+    display: inline-block;
+    position: sticky;
+    top: 96px; // gnb높이 만큼
+    line-height: 1.25;
+    max-height: 80vh;
+    overflow-y: auto;
+    word-break: keep-all;
+    border-left: 2px solid var(--color-text);
+    li {
+      margin: 0 0 0.6rem 0.6rem;
+      > ul {
+        margin: 0 0 0.6rem 0.6rem;
+      }
+    }
+    p {
+      margin: 0 0 0.6rem 0;
+      border-radius: 5px;
+    }
+    a {
+      display: inline-block;
+      &:hover {
+        color: var(--color-primary);
+      }
+    }
   }
+  @media (max-width: 1460px) {
+    right: -38%;
+  }
+
+  @media ${({ theme: { media } }) => media.xLarge} {
+    position: relative;
+    right: 0;
+    max-width: 100%;
+    > ul {
+      padding-left: 2rem;
+    }
+  }
+`
+
+// Footer
+
+export const Footer = styled.div`
+  width: var(--main-content-width);
+  margin: var(--padding-s) auto;
+  border-top: 1px solid var(--color-background-secondary);
+  padding-top: var(--padding-s);
 `
 
 // gatsby-plugin-image 라이브러리에서 제공해주는 GatsbyImage 컴포넌트에는 기본적으로 적용되어있는 인라인 스타일이 존재하는데,
@@ -105,7 +153,7 @@ export const BackgroundImage = styled((props: GatsbyImgProps) => (
   object-fit: cover;
   filter: brightness(0.25);
 
-  @media (max-width: 768px) {
+  @media ${({ theme }) => theme.media.medium} {
     height: 300px;
   }
 `
