--- src/components/PostDetail/PostFooter/PostFooter.style.ts
@@ -1,10 +1,0 @@
-import styled from '@emotion/styled'
-
-export const Container = styled.div`
-  width: 768px;
-  margin: 32px auto;
-  min-width: 680px;
-  border-top: 1px solid var(--color-background-secondary);
-  border-bottom: 1px solid var(--color-background-secondary);
-  padding-top: 32px;
-`
